refactor: create more general script used for displaying force diagram submission results
