--- Assets/Scripts/Activity 5/UI/AppleMotion/AppleForceDiagramSubmissionStatusDisplay.cs
@@ -1,18 +1,0 @@
-using System;
-
-public class AppleForceDiagramSubmissionStatusDisplay : ForceDiagramSubmissionStatusDisplay
-{
-	public static event Action ProceedEvent;
-
-	protected override void OnEnable()
-	{
-		base.OnEnable();
-		proceedButton.onClick.AddListener(() => ProceedEvent?.Invoke());
-	}
-
-	protected override void OnDisable()
-	{
-		base.OnDisable();
-		proceedButton.onClick.RemoveAllListeners();
-	}
-}
