--- Assets/Scripts/Activity 5/UI/AppleMotion/AppleForceDiagramSubmissionStatusDisplay.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6ea12ad6817c50c418843e84896d5e69
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
